note intégrée à Summoner et Summary + test sur getSummonerByName et getSummary fonctionnels

+++ src/main/java/com/project/ase_project/model/clean/grade/Grade.java
@@ -16,6 +16,6 @@ public class Grade {
     @Id
     private String id;
     private String summonerName;
-    private int average;
+    private float average;
     private int cardinal;
 }
